CB-10167 Create load balancers as part of Data Lake provisioning in Azure
For an Azure LB to be set up, it needs a few things:
 * The load balancer itself
 * A public IP address configuration
 * Backend pool (containing NICs to route to)
 * Health check rules
 * Forwarding rules

The LB SKU is Basic, because the NICs created for instances are Basic and they must be the same.
The Basic SKU supports up to 300 instances, does not support multiple AZs, and must route to
instances in the same Availability Set. The Basic SKU also does not have an SLA. The SKU associated
with NIC public IPs has been updated, along with load balancer and backend address pools.

The bulk of the changes are in the arm-v2.ftl file.
 * Addition of a loadbalancer resource
 * Addition of a publicIPAddress resource
 * Create a dependency between NICs in the gateway group and the LB
 * Place NICs that are part of the gateway group in the backend pool for the gateway load balancer

There are view objects that are used to pass relevant properties into the ARM template: AzureLoadBalancer,
AzureLoadBalancerProbe, and AzureLoadBalancingRule.

An implementation of collectLoadBalancer is provided in AzureMetadataCollector now which retrieves the
name, type (public/private), and ip address of load balancers.

Tested by setting up many, many medium duty datalakes while running CB locally. Set up and verified tear
down of a LIGHT_DUTY data lake as well.