CB-7972 Increase datalake start timeout to 120 min
There may be a situation where a 40 min timeout is not enough. For example if something is slow on the provider side, the polling could timeout on datalake side before the flow in the core could finish, thus the stack remains in UPDATE_IN_PROGRESS state. Increasing the timeout lets CB to finish the flow. Since we have flow tracking it should not cause any problems.